+++ VTK/IO/vtkImageReader2.cxx
@@ -141,7 +141,7 @@ void vtkImageReader2::ComputeInternalFileName(int slice)
     else if (this->FilePattern)
       {
       this->InternalFileName = new char [strlen(this->FilePattern) + 10];
-      int len = strlen(this->FilePattern);
+      int len = static_cast<int>(strlen(this->FilePattern));
       int hasPercentS = 0;
       for(int i =0; i < len-1; ++i)
         {
